requirements.txt

asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.3
idna==3.4
Pillow==10.0.1
protobuf==4.21.12
PyMySQL==1.1.0
requests==2.31.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.6
